Refactoring - MainManagerImpl: вынес повторяющийся код в новые методы (sortBooks, sortOrders); поправил в методах некоторые моменты, где можно сократить код и улучшить читаемость; - добавил проверки на корректность вводимых данных в getBookFromConsole(); - изменил в Book и Order типы полей price с Integer на Double, т.к. это вроде логичнее

+++ 4/View/OrdersMenu.java
@@ -22,5 +22,5 @@ public interface OrdersMenu extends Menu{
     void showErrorInputStatus();
 
     void showCountCompletedOrders(Long count);
-    void showEarnedSum(int sum);
+    void showEarnedSum(Double sum);
 }
